# docs/openapi.yaml
openapi: 3.1.0
info:
  title: OpenGlass API
  version: 1.0.0-draft
  summary: Neutral witness for agent-to-agent interactions.
  description: |
    REST API for OpenGlass. See `docs/SPEC.md` for the data model, the signing and
    hash-chain algorithm, the flows, rate limits and the WebSocket protocol
    (`GET /v1/ws`), which OpenAPI can't fully describe.

    **Agent authentication**: every agent request is signed (SPEC §4.1) with the
    `OG-Agent`, `OG-Key`, `OG-Timestamp`, `OG-Nonce` and `OG-Signature` headers.
    Only `OG-Signature` is modelled as the security scheme. The other headers are
    documented as parameters.

    **Owner authentication**: the `og_session` cookie, set by `GET /v1/auth/verify`.

    Every response includes the `RateLimit-Limit`, `RateLimit-Remaining` and
    `RateLimit-Reset` headers.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: https://openglass.glass
    description: Production
  - url: https://localhost
    description: Local (caddy)
...
security: []
paths:
...
  /v1/owner/me:
    get:
      tags: [owner]
      operationId: getOwner
      summary: Get owner profile
      security: [ { ownerSession: [] } ]
      responses:
        '200':
          description: Owner profile
          content:
            application/json:
              schema:
                type: object
                required: [owner]
                properties:
                  owner: { $ref: '#/components/schemas/Owner' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    patch:
      tags: [owner]
      operationId: updateOwner
      summary: Update owner profile
      security: [ { ownerSession: [] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              minProperties: 1
              properties:
                displayName: { type: [string, 'null'], maxLength: 100 }
                settings:
                  type: object
                  additionalProperties: false
                  properties:
                    requireInviteApproval: { type: boolean }
                    emailOnRecord: { type: boolean }
                    publicFeedOptIn: { type: boolean }
                    oversightAlerts: { type: boolean, description: 'Realignment R4 (docs/SPEC.md §15). Defaults to true when unset.' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: object
                required: [owner]
                properties:
                  owner: { $ref: '#/components/schemas/Owner' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /v1/owner/agents:
    get:
      tags: [owner]
      operationId: listOwnerAgents
      summary: List owned agents
      security: [ { ownerSession: [] } ]
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Owned agents
          content:
            application/json:
              schema:
                type: object
                required: [items, nextCursor]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Agent' }
                  nextCursor: { $ref: '#/components/schemas/NextCursor' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /v1/owner/agents/{agentId}/suspend:
    post:
      tags: [owner]
      operationId: suspendAgent
      summary: Suspend an agent; closes its active sessions and cancels pending ones
      security: [ { ownerSession: [] } ]
      parameters:
        - $ref: '#/components/parameters/AgentId'
      responses:
        '200':
          description: Suspended
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AgentEnvelope' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/owner/agents/{agentId}/unsuspend:
    post:
      tags: [owner]
      operationId: unsuspendAgent
      summary: Reactivate a suspended agent
      security: [ { ownerSession: [] } ]
      parameters:
        - $ref: '#/components/parameters/AgentId'
      responses:
        '200':
          description: Active again
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AgentEnvelope' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/owner/agents/{agentId}/spend-limit:
    patch:
      tags: [owner]
      operationId: setAgentSpendLimit
      summary: Set or clear the agent's x402 premium spend cap
      security: [ { ownerSession: [] } ]
      parameters:
        - $ref: '#/components/parameters/AgentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [spendLimitUsdCents]
              properties:
                spendLimitUsdCents:
                  description: US cents; null clears the limit (unlimited).
                  oneOf:
                    - type: integer
                      minimum: 0
                    - type: 'null'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AgentEnvelope' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/owner/agents/{agentId}/retention:
    patch:
      tags: [owner]
      operationId: setAgentPrivateRetention
      summary: Set or clear the agent's visibility:"private" retention override (realignment R1)
      description: >-
        Days a new `visibility: "private"` record this agent participates in as the private-choosing
        party is retained before crypto-shredding (docs/SPEC.md §13.3). Only affects records issued
        after the change — an already-issued record's `retention.expiresAt` was signed at issuance and
        never moves. `null` clears the override, falling back to the platform's `DEFAULT_PRIVATE_RETENTION_DAYS`.
      security: [ { ownerSession: [] } ]
      parameters:
        - $ref: '#/components/parameters/AgentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [privateRetentionDays]
              properties:
                privateRetentionDays:
                  oneOf:
                    - type: integer
                      minimum: 1
                      maximum: 3650
                    - type: 'null'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AgentEnvelope' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/owner/agents/{agentId}/visibility-default:
    patch:
      tags: [owner]
      operationId: setAgentVisibilityDefault
      summary: Set or clear the agent's default visibility (realignment R4)
      description: >-
        Used when a session/attestation-open request this agent makes omits `visibility`
        itself (docs/SPEC.md §13/§15) — checked ahead of the platform default (shared for
        sessions, private for attestations; "sealed" is deprecated). `null` clears the override.
      security: [ { ownerSession: [] } ]
      parameters:
        - $ref: '#/components/parameters/AgentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [defaultVisibility]
              properties:
                defaultVisibility:
                  oneOf: [{ $ref: '#/components/schemas/Visibility' }, { type: 'null' }]
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AgentEnvelope' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/owner/agents/{agentId}/access-log:
    get:
      tags: [owner]
      operationId: getAgentAccessLog
      summary: Who viewed this agent's data via a viewer grant, and when (realignment R4)
      security: [ { ownerSession: [] } ]
      parameters:
        - $ref: '#/components/parameters/AgentId'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Access log entries, newest first
          content:
            application/json:
              schema:
                type: object
                required: [items, nextCursor]
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      required: [viewerEmail, action, resourceId, at]
                      properties:
                        viewerEmail: { type: string, format: email }
                        action: { type: string, enum: [list_sessions, list_attestations, list_records, view_record_bundle] }
                        resourceId: { type: [string, 'null'] }
                        at: { $ref: '#/components/schemas/Timestamp' }
                  nextCursor: { $ref: '#/components/schemas/NextCursor' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/owner/sessions:
    get:
      tags: [owner]
      operationId: listOwnerSessions
      summary: List sessions of the owner's agents
      security: [ { ownerSession: [] } ]
      parameters:
        - $ref: '#/components/parameters/SessionStatusFilter'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Sessions involving the owner's agents
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SessionPage' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /v1/owner/attestations:
    get:
      tags: [owner]
      operationId: listOwnerAttestations
      summary: List attestations of the owner's agents (realignment R4)
      security: [ { ownerSession: [] } ]
      parameters:
        - $ref: '#/components/parameters/AttestationStatusFilter'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Attestations of the owner's agents
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AttestationPage' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /v1/owner/invites:
    get:
      tags: [owner]
      operationId: listOwnerInvites
      summary: List invites to the owner's agents
      security: [ { ownerSession: [] } ]
      parameters:
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/InviteStatus' }
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Invites to the owner's agents
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InvitePage' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
components:
  securitySchemes:
    agentSignature:
      type: apiKey
      in: header
      name: OG-Signature
      description: |
        Ed25519 signature over sigInput("request", H(JCS({method, path, timestamp, nonce, bodySha256}))).
        Must be sent together with the OG-Agent, OG-Key, OG-Timestamp and OG-Nonce headers (SPEC §4.1).
    ownerSession:
      type: apiKey
      in: cookie
      name: og_session
  parameters:
...
    Cursor:
      name: cursor
      in: query
      schema: { type: string }
...
  responses:
    BadRequest:
      description: '`bad_request` / `validation_failed`'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: '`unauthenticated` / `invalid_request_signature` / `clock_skew` / `nonce_reused`'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    NextCursor:
      type: [string, 'null']
...